panel/panel/requirements/production.txt
django==1.9
djangorestframework
pyyaml
pytest-django
celery
redis
pycrypto
